refactor lambda runner into own class
