+++ package.json
@@ -23,6 +23,7 @@
     "body-parser": "^1.18.3",
     "cookie-parser": "^1.4.3",
     "express": "^4.16.3",
+    "js-yaml": "^3.12.0",
     "yargs": "^11.0.0"
   },
   "peerDependencies": {
