+++ test/00-require.js
@@ -1,7 +1,7 @@
 const assert = require('assert'),
   path = require('path');
 
-const files = ['utils.js', 'index.js'].map(f => path.resolve(f, '..'));
+const files = ['lib/lambda', 'utils', 'index'].map(f => path.resolve(f));
 
 describe('require', () => {
   files.map(file => {
