+++ utils.js
@@ -1,3 +1,17 @@
 module.exports = {
-
+  ENV_VARS: [
+    'AWS_REGION',
+    'AWS_DEFAULT_REGION',
+    'AWS_ACCOUNT_ID',
+    'AWS_ACCESS_KEY_ID',
+    'AWS_SECRET_ACCESS_KEY',
+    'AWS_SESSION_TOKEN',
+    'AWS_LAMBDA_FUNCTION_NAME',
+    'AWS_LAMBDA_FUNCTION_VERSION',
+    'AWS_LAMBDA_FUNCTION_MEMORY_SIZE',
+    'AWS_LAMBDA_FUNCTION_TIMEOUT',
+    'AWS_LAMBDA_FUNCTION_HANDLER',
+    'AWS_LAMBDA_EVENT_BODY',
+    'DOCKER_LAMBDA_USE_STDIN',
+  ],
 };
